requirements.txt
# torch
# torchaudio
# torchvision
transformers
datasets[audio]
webdataset
accelerate
einops
hydra-core
tensorboardx